RavenDB-2687 Creating index with incorrect custom analyzer name creates the index and breaks database
